фикс запроса баланса

+++ internal/domain/models/balance.go
@@ -2,5 +2,5 @@ package models
 
 type Balance struct {
 	Current   float64 `json:"current"`
-	Withdrawn int64   `json:"withdrawn"`
+	Withdrawn float64 `json:"withdrawn"`
 }
